# swagger_final.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MilkApiManager",
    "version": "1.0"
  },
  "paths": {
...
    "/api/PiiMasking": {
      "get": {
        "tags": [
          "PiiMasking"
        ],
        "summary": "Retrieves all PII masking rules.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PiiMaskingRule"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PiiMaskingRule"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PiiMaskingRule"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PiiMasking"
        ],
        "summary": "Creates a new PII masking rule governing APISIX traffic.",
        "requestBody": {
          "description": "The rule definition.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PiiMaskingRule"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PiiMaskingRule"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PiiMaskingRule"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PiiMaskingRule"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PiiMaskingRule"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PiiMaskingRule"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/PiiMasking/{id}": {
      "put": {
        "tags": [
          "PiiMasking"
        ],
        "summary": "Updates an existing PII masking rule.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The rule ID to update.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "The new rule content.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PiiMaskingRule"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PiiMaskingRule"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PiiMaskingRule"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PiiMasking"
        ],
        "summary": "Deletes a PII masking rule from the system and APISIX.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The rule ID to delete.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PiiMaskingRule": {
        "required": [
          "fieldPath",
          "regexPattern",
          "replacePattern",
          "routeId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "routeId": {
            "minLength": 1,
            "type": "string"
          },
          "fieldPath": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "regexPattern": {
            "minLength": 1,
            "type": "string"
          },
          "replacePattern": {
            "minLength": 1,
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}